paperlens-multimodal-auditor/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-genai==0.3.0
pymupdf==1.23.8
numpy==1.26.4
pillow==10.1.0
python-multipart==0.0.6
